Restructure playwright tests
