+++ .github/workflows/ci-cd.yml
@@ -34,7 +34,7 @@ jobs:
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Run tests
-      run: pytest --ignore=tests/playwright/test_snapshot.py
+      run: pytest -m "not snapshot"
 
   check-javascript:
     runs-on: ubuntu-latest
